+++ test/sync/src/commonMain/kotlin/io/realm/kotlin/entities/sync/SyncObjectWithAllTypes.kt
@@ -17,11 +17,13 @@
 package io.realm.kotlin.entities.sync
 
 import io.realm.kotlin.ext.realmListOf
+import io.realm.kotlin.ext.realmSetOf
 import io.realm.kotlin.schema.RealmStorageType
 import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmSet
 import io.realm.kotlin.types.RealmUUID
 import io.realm.kotlin.types.annotations.PrimaryKey
 import kotlin.random.Random
@@ -84,6 +86,25 @@ class SyncObjectWithAllTypes : RealmObject {
     // Nullable RealmLists of primitive values, not currently supported by Sync
     // Nullable Object lists, not currently supported by Core
 
+    // RealmSets
+    var stringRealmSet: RealmSet<String> = realmSetOf("hello world")
+    var byteRealmSet: RealmSet<Byte> = realmSetOf(0)
+    var charRealmSet: RealmSet<Char> = realmSetOf(0.toChar())
+    var shortRealmSet: RealmSet<Short> = realmSetOf(0)
+    var intRealmSet: RealmSet<Int> = realmSetOf(0)
+    var longRealmSet: RealmSet<Long> = realmSetOf(0)
+    var booleanRealmSet: RealmSet<Boolean> = realmSetOf(true)
+    var doubleRealmSet: RealmSet<Double> = realmSetOf(0.0)
+    var floatRealmSet: RealmSet<Float> = realmSetOf(0.0.toFloat())
+    var realmInstantRealmSet: RealmSet<RealmInstant> = realmSetOf(RealmInstant.MIN)
+    var objectIdRealmSet: RealmSet<ObjectId> = realmSetOf(ObjectId.create())
+    var realmUUIDRealmSet: RealmSet<RealmUUID> = realmSetOf(RealmUUID.random())
+    var binaryRealmSet: RealmSet<ByteArray> = realmSetOf(byteArrayOf(42))
+    var objectRealmSet: RealmSet<SyncObjectWithAllTypes> = realmSetOf()
+
+    // RealmSets of nullable primitive values, not currently supported by Sync
+    // RealmSets of nullable objects, not currently supported by Core
+
     companion object {
 
         // Mapping between each Core Field type and functions that can insert data for that type
@@ -100,11 +121,13 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.intField = 42
                                         obj.intNullableField = 42
                                         obj.intRealmList = realmListOf(42)
+                                        obj.intRealmSet = realmSetOf(42)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(42, obj.intField)
                                         assertEquals(42, obj.intNullableField)
                                         assertEquals(42, obj.intRealmList.first())
+                                        assertSetContains(42, obj.intRealmSet)
                                     }
                                 )
                             }
@@ -114,12 +137,15 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.booleanField = true
                                         obj.booleanNullableField = true
                                         obj.booleanRealmList = realmListOf(true, false)
+                                        obj.booleanRealmSet = realmSetOf(true, false)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(true, obj.booleanField)
                                         assertEquals(true, obj.booleanNullableField)
                                         assertEquals(true, obj.booleanRealmList[0])
                                         assertEquals(false, obj.booleanRealmList[1])
+                                        assertSetContains(true, obj.booleanRealmSet)
+                                        assertSetContains(false, obj.booleanRealmSet)
                                     }
                                 )
                             }
@@ -129,12 +155,15 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.stringField = "Foo"
                                         obj.stringNullableField = "Bar"
                                         obj.stringRealmList = realmListOf("Foo", "")
+                                        obj.stringRealmSet = realmSetOf("Foo", "")
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals("Foo", obj.stringField)
                                         assertEquals("Bar", obj.stringNullableField)
                                         assertEquals("Foo", obj.stringRealmList[0])
                                         assertEquals("", obj.stringRealmList[1])
+                                        assertSetContains("Foo", obj.stringRealmSet)
+                                        assertSetContains("", obj.stringRealmSet)
                                     }
                                 )
                             }
@@ -151,6 +180,12 @@ class SyncObjectWithAllTypes : RealmObject {
                                                     stringField = "child2"
                                                 }
                                             )
+                                        obj.objectRealmSet =
+                                            realmSetOf(
+                                                SyncObjectWithAllTypes().apply {
+                                                    stringField = "child2"
+                                                }
+                                            )
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals("child1", obj.objectField!!.stringField)
@@ -159,6 +194,7 @@ class SyncObjectWithAllTypes : RealmObject {
                                             "child2",
                                             obj.objectRealmList.first().stringField
                                         )
+                                        assertSetContainsObject("child2", obj.objectRealmSet)
                                     }
                                 )
                             }
@@ -169,6 +205,8 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.floatNullableField = 1.23F
                                         obj.floatRealmList =
                                             realmListOf(1.23F, Float.MIN_VALUE, Float.MAX_VALUE)
+                                        obj.floatRealmSet =
+                                            realmSetOf(1.23F, Float.MIN_VALUE, Float.MAX_VALUE)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(1.23F, obj.floatField)
@@ -176,6 +214,8 @@ class SyncObjectWithAllTypes : RealmObject {
                                         assertEquals(1.23F, obj.floatRealmList[0])
                                         assertEquals(Float.MIN_VALUE, obj.floatRealmList[1])
                                         assertEquals(Float.MAX_VALUE, obj.floatRealmList[2])
+                                        assertSetContains(Float.MIN_VALUE, obj.floatRealmSet)
+                                        assertSetContains(Float.MAX_VALUE, obj.floatRealmSet)
                                     }
                                 )
                             }
@@ -186,6 +226,8 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.doubleNullableField = 1.234
                                         obj.doubleRealmList =
                                             realmListOf(1.234, Double.MIN_VALUE, Double.MAX_VALUE)
+                                        obj.doubleRealmSet =
+                                            realmSetOf(1.234, Double.MIN_VALUE, Double.MAX_VALUE)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(1.234, obj.doubleField)
@@ -193,6 +235,8 @@ class SyncObjectWithAllTypes : RealmObject {
                                         assertEquals(1.234, obj.doubleRealmList[0])
                                         assertEquals(Double.MIN_VALUE, obj.doubleRealmList[1])
                                         assertEquals(Double.MAX_VALUE, obj.doubleRealmList[2])
+                                        assertSetContains(Double.MIN_VALUE, obj.doubleRealmSet)
+                                        assertSetContains(Double.MAX_VALUE, obj.doubleRealmSet)
                                     },
                                 )
                             }
@@ -204,6 +248,8 @@ class SyncObjectWithAllTypes : RealmObject {
                                             RealmInstant.from(-1, -1)
                                         obj.realmInstantRealmList =
                                             realmListOf(RealmInstant.MIN, RealmInstant.MAX)
+                                        obj.realmInstantRealmSet =
+                                            realmSetOf(RealmInstant.MIN, RealmInstant.MAX)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(
@@ -216,6 +262,14 @@ class SyncObjectWithAllTypes : RealmObject {
                                         )
                                         assertEquals(RealmInstant.MIN, obj.realmInstantRealmList[0])
                                         assertEquals(RealmInstant.MAX, obj.realmInstantRealmList[1])
+                                        assertSetContains(
+                                            RealmInstant.MIN,
+                                            obj.realmInstantRealmSet
+                                        )
+                                        assertSetContains(
+                                            RealmInstant.MAX,
+                                            obj.realmInstantRealmSet
+                                        )
                                     },
                                 )
                             }
@@ -228,12 +282,15 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.objectIdField = randomObjId
                                         obj.objectIdNullableField = randomObjId
                                         obj.objectIdRealmList = realmListOf(minObjId, maxObjId)
+                                        obj.objectIdRealmSet = realmSetOf(minObjId, maxObjId)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(randomObjId, obj.objectIdField)
                                         assertEquals(randomObjId, obj.objectIdNullableField)
                                         assertEquals(minObjId, obj.objectIdRealmList[0])
                                         assertEquals(maxObjId, obj.objectIdRealmList[1])
+                                        assertSetContains(minObjId, obj.objectIdRealmSet)
+                                        assertSetContains(maxObjId, obj.objectIdRealmSet)
                                     },
                                 )
                             }
@@ -246,12 +303,15 @@ class SyncObjectWithAllTypes : RealmObject {
                                         obj.realmUUIDField = uuid1
                                         obj.realmUUIDNullableField = uuid1
                                         obj.realmUUIDRealmList = realmListOf(uuid2, uuid3)
+                                        obj.realmUUIDRealmSet = realmSetOf(uuid2, uuid3)
                                     },
                                     { obj: SyncObjectWithAllTypes ->
                                         assertEquals(uuid1, obj.realmUUIDField)
                                         assertEquals(uuid1, obj.realmUUIDNullableField)
                                         assertEquals(uuid2, obj.realmUUIDRealmList[0])
                                         assertEquals(uuid3, obj.realmUUIDRealmList[1])
+                                        assertSetContains(uuid2, obj.realmUUIDRealmSet)
+                                        assertSetContains(uuid3, obj.realmUUIDRealmSet)
                                     },
                                 )
                             }
@@ -260,19 +320,50 @@ class SyncObjectWithAllTypes : RealmObject {
                                     { obj: SyncObjectWithAllTypes ->
                                         obj.binaryField = byteArrayOf(22)
                                         obj.binaryNullableField = byteArrayOf(22)
-                                        obj.binaryRealmList =
-                                            realmListOf(
-                                                byteArrayOf(22),
-                                                byteArrayOf(44, 66),
-                                                byteArrayOf(11, 33)
-                                            )
+                                        obj.binaryRealmList = realmListOf(
+                                            byteArrayOf(22),
+                                            byteArrayOf(44, 66),
+                                            byteArrayOf(11, 33)
+                                        )
+                                        obj.binaryRealmSet = realmSetOf(
+                                            byteArrayOf(22),
+                                            byteArrayOf(44, 66),
+                                            byteArrayOf(11, 33)
+                                        )
                                     },
                                     { obj: SyncObjectWithAllTypes ->
-                                        assertContentEquals(byteArrayOf(22), obj.binaryField)
-                                        assertContentEquals(byteArrayOf(22), obj.binaryNullableField)
-                                        assertContentEquals(byteArrayOf(22), obj.binaryRealmList[0])
-                                        assertContentEquals(byteArrayOf(44, 66), obj.binaryRealmList[1])
-                                        assertContentEquals(byteArrayOf(11, 33), obj.binaryRealmList[2])
+                                        assertContentEquals(
+                                            byteArrayOf(22),
+                                            obj.binaryField
+                                        )
+                                        assertContentEquals(
+                                            byteArrayOf(22),
+                                            obj.binaryNullableField
+                                        )
+                                        assertContentEquals(
+                                            byteArrayOf(22),
+                                            obj.binaryRealmList[0]
+                                        )
+                                        assertContentEquals(
+                                            byteArrayOf(44, 66),
+                                            obj.binaryRealmList[1]
+                                        )
+                                        assertContentEquals(
+                                            byteArrayOf(11, 33),
+                                            obj.binaryRealmList[2]
+                                        )
+                                        assertSetContainsBinary(
+                                            byteArrayOf(22),
+                                            obj.binaryRealmSet
+                                        )
+                                        assertSetContainsBinary(
+                                            byteArrayOf(44, 66),
+                                            obj.binaryRealmSet
+                                        )
+                                        assertSetContainsBinary(
+                                            byteArrayOf(11, 33),
+                                            obj.binaryRealmSet
+                                        )
                                     },
                                 )
                             }
@@ -287,11 +378,47 @@ class SyncObjectWithAllTypes : RealmObject {
             }
         }
 
+        private fun assertSetContains(value: Any?, set: RealmSet<*>) {
+            if (!set.contains(value)) {
+                throw IllegalStateException("Set doesn't contain value $value")
+            }
+        }
+
+        // Sets don't expose indices so we need to iterate them
+        private fun assertSetContainsObject(value: String, set: RealmSet<SyncObjectWithAllTypes>) {
+            var found = false
+            val iterator = set.iterator()
+            while (iterator.hasNext()) {
+                val obj = iterator.next()
+                if (obj.stringField == value) {
+                    found = true
+                }
+            }
+            if (!found) {
+                throw IllegalStateException("Set doesn't contain object with 'stringField' value '$value'")
+            }
+        }
+
+        // Similarly we need to iterate over the set and see if the binary contests are the same
+        private fun assertSetContainsBinary(value: ByteArray, set: RealmSet<ByteArray>) {
+            val iterator = set.iterator()
+            var found = false
+            while (iterator.hasNext()) {
+                val byteArray = iterator.next()
+                if (value.contentEquals(byteArray)) {
+                    found = true
+                }
+            }
+            if (!found) {
+                throw IllegalStateException("Set does not contain ByteArray $value")
+            }
+        }
+
         private fun assertContentEquals(value: ByteArray?, other: ByteArray?) {
             value?.forEachIndexed { index, byte ->
                 val actual = other?.get(index)
                 if (byte != actual) {
-                    throw IllegalStateException("Values do not match: '$value' vs. '$other'")
+                    throw IllegalStateException("Values do not match: '$byte' vs. '$actual'")
                 }
             }
         }